+++ src/lib/site-urls.js
@@ -1,0 +1,5 @@
+export const NPM_URL = 'https://www.npmjs.com/package/md-mermaid-pdf'
+export const GITHUB_LIB_URL = 'https://github.com/Ali-Karaki/md-mermaid-pdf'
+export const GITHUB_SITE_URL = 'https://github.com/Ali-Karaki/md-mermaid-pdf-site'
+export const MD_TO_PDF_OPTIONS =
+  'https://github.com/simonhaenisch/md-to-pdf#options'